requirements.txt
# Runtime dependencies — all other imports are Python stdlib.
# Install with: pip install -r requirements.txt
anthropic>=0.25.0
rich>=13.0.0
python-dotenv>=1.0.0
